requirements.txt
annotated-types==0.7.0
anyio==4.4.0
astmonkey==0.3.6
attrs==23.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
cryptography==43.0.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.8
einops==0.8.0
emoji==2.12.1
exceptiongroup==1.2.2
fastapi==0.112.2
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
libretranslatepy==2.1.1
loguru==0.7.2
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
memory-profiler==0.61.0
mpmath==1.3.0
murmurhash==1.0.10
MutPy==0.6.1
mysql-connector-python==9.0.0
networkx==3.2.1
nltk==3.9.1
NRCLex==3.0.0
numpy==1.26.4
ollama==0.3.1
outcome==1.3.0.post0
packaging==24.1
preshed==3.0.9
protobuf==5.28.0
psutil==6.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydot==3.0.1
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.1.4
pyphen==0.16.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.8.0
sacremoses==0.1.1
selenium==4.24.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.7.6
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacytextblob==4.0.0
SQLAlchemy==2.0.32
srsly==2.4.8
stanfordcorenlp==3.9.1.1
stanza==1.8.2
starlette==0.38.4
sympy==1.13.2
termcolor==2.4.0
textblob==0.15.3
textstat==0.7.4
thinc==8.2.5
tiktoken==0.7.0
toml==0.10.2
torch==2.4.0
tqdm==4.66.5
translate==3.6.1
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.2.2
uvicorn==0.30.6
vaderSentiment==3.3.2
validate-email-address==1
validators==0.34.0
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
websockets==13.0.1
wrapt==1.16.0
wsproto==1.2.0
